% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gdh.R
\name{calc_gdh}
\alias{calc_gdh}
\title{Calculate Growing Degree Hours}
\usage{
calc_gdh(data, field = NULL)
}
\arguments{
\item{data}{is the dataframe that is output from the 'combine_df' function}
}
\value{
the growing degree hours for the dataframe provided
}
\description{
Calculate Growing Degree Hours
}
\examples{
data = temp_df
growing_degrees <- calc_gdh(data)
print(growing_degrees)
}
