% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hourly_temp.R
\name{calc_hourly_temp}
\alias{calc_hourly_temp}
\title{Calculate hourly temperature}
\usage{
calc_hourly_temp(data)
}
\arguments{
\item{data}{is the dataframe that contains the temp data. Cols should be named 'date_time', 'air_temp_set_1'}
}
\description{
Calculate hourly temperature
}
