% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_df.R
\name{combine_df}
\alias{combine_df}
\title{Create Temporary Dataframe to Calculate GDHs}
\usage{
combine_df(temp_data, database_data, gdh_data)
}
\arguments{
\item{temp_data}{is field_weather_data from 'create_field_weather_data'}

\item{database_data}{is the field information, including crop and seeding date}

\item{gdh_data}{is the information needed to calculate GDHs, including base and upper temps}
}
\description{
Create Temporary Dataframe to Calculate GDHs
}
