% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_field_weather_data.R
\name{create_field_weather_data}
\alias{create_field_weather_data}
\title{Summary of temperature aggregations}
\usage{
create_field_weather_data(data)
}
\arguments{
\item{data}{is the dataframe that contains the temperatures from each station. Temp column should be named "temp_avg}
}
\description{
Summary of temperature aggregations
}
