% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_crops.R
\name{download_crops}
\alias{download_crops}
\title{Request crop data (seed_harvest_wide) from Snowflake within a certain season range}
\usage{
download_crops(conn, gdd_info_formatted, min_year = NA, max_year = NA)
}
\arguments{
\item{conn}{The connection to the Snowflake database}

\item{min_year}{Minimum crop season year to pull from. Defaults to NA (pull all years)}

\item{max_year}{Maximum crop season year to pull from. Defaults to NA (pull all years)}
}
\value{
The seed_harvest_df data frame.
}
\description{
Request crop data (seed_harvest_wide) from Snowflake within a certain season range
}
\examples{
seed_harvest_wide <- download_crops(myconn, min_year=2020, max_year=2022)
}
