% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{elev_change}
\alias{elev_change}
\title{Calculate elevation difference}
\usage{
elev_change(data)
}
\arguments{
\item{data}{is the dataframe that contains the elevations. Columns should be named "field_elev" and "station_elev"}
}
\description{
Calculate elevation difference
}
