% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closest_stations.R
\name{find_closest_stations}
\alias{find_closest_stations}
\title{Find the closest stations}
\usage{
find_closest_stations(field_df, stations_df, radius)
}
\arguments{
\item{field_df}{is the fields dataset.}

\item{stations_df}{is the dataset that includes all the active stations that was derived from the pull_active_stations() function.}

\item{radius}{is set to the maximum distance a station could be away from a field.}
}
\description{
Find the closest stations
}
\examples{
p <- find_closest_stations(field_df, stations_df, radius = ###)
}
