% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.r
\name{get_all_fields}
\alias{get_all_fields}
\title{Get all fields from the Snowflake database}
\usage{
get_all_fields(myconn)
}
\arguments{
\item{myconn}{The connection to the Snowflake database}
}
\description{
Get all fields from the Snowflake database
}
\examples{
all_fields1 <- get_all_fields(myconn)
}
