% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get_elev}
\alias{get_elev}
\title{Get Elevations}
\usage{
get_elev(df, conn = NA)
}
\arguments{
\item{df}{is the dataframe as output from the 'all_fields' function}
}
\description{
Get Elevations
}
\examples{
 df = makeaneexample
get_elev(df)

}
