% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.r
\name{get_latest_date}
\alias{get_latest_date}
\title{Get latest date with GDHs from the Snowflake database}
\usage{
get_latest_date(myconn)
}
\arguments{
\item{myconn}{The connection to the Snowflake database}
}
\description{
Get latest date with GDHs from the Snowflake database
}
\examples{
latest_date <- get_latest_date(myconn)
}
