% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.r
\name{get_seed_harvest_wide}
\alias{get_seed_harvest_wide}
\title{Get the wide seed harvest dataframe (based on field crops data) from the Snowflake database}
\usage{
get_seed_harvest_wide(myconn, gdd_info_formatted)
}
\arguments{
\item{myconn}{The connection to the Snowflake database}

\item{gdd_info_formatted}{Growing Degree Day temperature threshold data}
}
\description{
Get the wide seed harvest dataframe (based on field crops data) from the Snowflake database
}
\examples{
seed_harvest_df <- get_seed_harvest(myconn)
}
