% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get_state}
\alias{get_state}
\title{Get state abbreviations}
\usage{
get_state(df)
}
\arguments{
\item{df}{is the dataframe that includes the coordinate columns. Make sure to rename columns to 'lat', 'lon'}
}
\description{
Get state abbreviations
}
