% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_temp_all.R
\docType{data}
\name{global_resp}
\alias{global_resp}
\title{Request the temperature data from MesoWest, but fast}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
global_resp
}
\arguments{
\item{df}{is the fields_stations dataset derived from the find_closest_station() function.}
}
\description{
Request the temperature data from MesoWest, but fast
}
\examples{
p <- pull_temp_all(df)
}
\keyword{datasets}
