% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{lat_dist}
\alias{lat_dist}
\title{Calculate latitudinal distance}
\usage{
lat_dist(data)
}
\arguments{
\item{data}{is the dataframe that contains the latitude. Column should be named "lat"}
}
\description{
Calculate latitudinal distance
}
\examples{
 data 
p <- lat_dist(data)
}
