% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{lon_dist}
\alias{lon_dist}
\title{Calculate longitudinal distance}
\usage{
lon_dist(data)
}
\arguments{
\item{data}{is the dataframe that contains the latitude. Column should be named "lon"}
}
\description{
Calculate longitudinal distance
}
\examples{
data 
p <- lon_dist(data)
}
