% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_active_stations.R
\name{pull_active_stations}
\alias{pull_active_stations}
\title{Pull active stations from MesoWest API}
\usage{
pull_active_stations(df, key)
}
\arguments{
\item{df}{is 'fields' dataset, including FIELD id,lat,lon,elevation,state, CROP name,season, seeding and harvest dates}

\item{key}{is the API key for MesoWest}
}
\description{
Pull active stations from MesoWest API
}
\examples{
p <- pull_active_stations(df, key)
}
