% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_temp.R
\name{pull_temp_all}
\alias{pull_temp_all}
\title{Request the temperature data from MesoWest quickly}
\usage{
pull_temp_all(df)
}
\arguments{
\item{df}{is the fields_stations dataset derived from the find_closest_station() function.}
}
\description{
Request the temperature data from MesoWest
}
\examples{
p <- pull_temp_all(df)
}
