% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_temp.R
\name{pull_temp_full}
\alias{pull_temp_full}
\title{Request the hourly temperature data from MesoWest and process it}
\usage{
pull_temp_full(fields_stations)
}
\arguments{
\item{fields_stations}{is the fields_stations dataset derived from the find_closest_station() function.}
}
\description{
Request the hourly temperature data from MesoWest and process it
}
\examples{
station_temp_data <- pull_temp_full(fields_stations)
}
