% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.r
\name{push_gdh}
\alias{push_gdh}
\title{Push GDH data to Snowflake.}
\usage{
push_gdh(myconn, table = "GDH_BYUI_DEV", growing_degree_units)
}
\arguments{
\item{myconn}{A connection to the Snowflake database.}

\item{table}{The table to push to.}

\item{growing_degree_units}{The GDH data to push.}
}
\description{
Push GDH data to Snowflake.
}
