% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.r
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicate rows from a Snowflake table}
\usage{
remove_duplicates(myconn, table)
}
\arguments{
\item{myconn}{A connection to the Snowflake database}

\item{table}{The table to remove duplicates from}
}
\description{
Remove duplicate rows from a Snowflake table
}
\examples{
remove_duplicates(table)
}
