% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{total_dist}
\alias{total_dist}
\title{Calculate distance}
\usage{
total_dist(data)
}
\arguments{
\item{data}{is the dataframe that contains the latitude and longitude. Columns should be named "lat" and "lon"}
}
\description{
Calculate distance
}
\examples{
data 
p <- total_dist(data)
}
