% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{weighted}
\alias{weighted}
\title{Calculate weights according to elevation}
\usage{
weighted(data)
}
\arguments{
\item{data}{is the dataframe that contains the elevations. Columns should be named "FIELD_ID" and "station_id"}
}
\description{
Calculate weights according to elevation
}
