% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gdh.R
\name{calc_gdh}
\alias{calc_gdh}
\title{Calculate Growing Degree Hours}
\usage{
calc_gdh(data, field = NULL)
}
\arguments{
\item{data}{is the dataframe that is output from 'combine_df' function}
}
\description{
Calculate Growing Degree Hours
}
